+++ devices/sensor/barometer.py
@@ -7,7 +7,26 @@ class BarometerSensor(Device):
         return Domoticz.Device(Unit=unit, DeviceID=device_id, Name=device_name, TypeName="Barometer").Create()
 
     def get_numeric_value(self, value, device):
-        return int(value)
+        return 0
 
     def get_string_value(self, value, device):
-        return '0'
+        return ';'.join([
+            str(value),
+            str(self.get_forecast(value))
+        ])
+
+    def get_forecast(self, value):
+        # Forecast is based on https://github.com/stas-demydiuk/domoticz-zigbee2mqtt-plugin/issues/102
+
+        if (value < 996):
+            return 4  # Thunderstorm
+        elif (value < 993):
+            return 3  # Not stable
+        elif (value < 1007):
+            return 2  # Cloudy/Rain
+        elif (value < 1013):
+            return 1  # Clear/Sunny
+        elif (value < 1033):
+            return 0  # Stable
+        else:
+            return 5  # Unknown