+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## master
 - Add temperature sensor support for Xiaomi Mi power plug
 - Add barometer sensor for Xiaomi Aqara temperature, humidity and pressure sensor (WSDCGQ11LM)
+- Add forecast support for WSDCGQ11LM
 
 New devices:
 - Bosch Motion sensor